app/requirements.txt
django
aiofiles==0.7.0
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0  
idna==2.10
mac-vendor-lookup==0.1.11
multidict==5.1.0
numpy==1.20.3
opencv-python
python-nmap==0.6.4
pytz==2021.1
requests==2.25.1
scapy==2.4.5
simple-http-server==0.11.4
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
netifaces
vulners==1.5.11
pymongo[snappy,gssapi,srv,tls
dnspython



# mac-vendor-lookup
# scapy
# opencv-python
# netifaces-0.11.0
# netifaces


# netifaces==0.11.0
